Version 0.3.0
This is a very large and is a rewrite of the entire Nation part of the wrapper. It is incompatible with all prior programs who used the part, and the changes are as such:

- Fixed the dump collection functions, will now unzip the data to return usable String.
- Added issue type for usability purposes, it is returned by the issues property
- Made property access syntax for all nation shards.
- Remade how commands work and added specific functions for issues, dispatches and giftcards.
- Changed UserAgent to be more informative.
- Updated REAMDE and fixed typos.
- Removed ambiguous return type functions
- Changed library license in the setup
- Added required libraries to setup
- Functions now use the APICall UserAgent.

I intend to rework some other parts and make them less xml based (such as the TradingCards part), but this will be done more gradually and with deprecation notices, and will likely be the 0.3.x updates.

I have been working on this for a while now, and I hope this will be a positive change